Move useAssetLayer and related functions one level up
