+++ apps/melding-form/src/app/(map)/locatie/kies/_components/Map/Map.module.css
@@ -12,8 +12,3 @@
 .hideMap {
   display: none;
 }
-
-.selectedMarkerIconBorder {
-  border: 4px solid var(--ams-color-highlight-green); /* stylelint-disable-line unit-disallowed-list */
-  border-radius: 50%;
-}
