+++ apps/melding-form/src/app/(map)/locatie/kies/_utils/index.ts
@@ -1,1 +1,4 @@
 export { getAddressFromCoordinates } from './getAddressFromCoordinates'
+export { addAssetLayerToMap } from './addAssetLayerToMap'
+export { ASSET_ZOOM_THRESHOLD, fetchAssets } from './fetchAssets'
+export { updateAssetMarkers } from './updateAssetMarkers'
