+++ apps/melding-form/src/app/(map)/locatie/kies/_utils/updateAssetMarkers.test.ts
@@ -9,7 +9,7 @@ vi.mock('../markerIcons', () => ({
   selectedAssetsIcon: { icon: 'selected' },
 }))
 
-vi.mock('../utils/getContainerFeatureIcon', () => ({
+vi.mock('./getContainerFeatureIcon', () => ({
   getContainerFeatureIcon: vi.fn((_feature, isSelected) => ({ icon: `container${isSelected ? '-selected' : ''}` })),
 }))
